backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

SQLAlchemy==2.0.36
asyncpg==0.29.0

pydantic==2.9.0
pydantic-settings==2.4.0

passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-jose[cryptography]==3.3.0
email-validator>=2.0.0
python-multipart==0.0.9